requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2.2
django-phonenumber-field==5.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
environs==9.3.2
filelock==3.0.12
idna==2.10
marshmallow==3.11.1
phonenumbers==8.12.22
phonenumberslite==8.12.22
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
structlog==21.1.0
swapper==1.1.2.post1
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
